+++ clients/vscode-hlasmplugin/CHANGELOG.md
@@ -8,6 +8,9 @@
 - The language server crashes when a redefined data constant has a dependency
 - Infinite loop when an incomplete data definition is found in the lookahead mode
 
+#### Changed
+- New parser implementation
+
 ## [1.15.1](https://github.com/eclipse-che4z/che-che4z-lsp-for-hlasm/compare/1.15.0...1.15.1) (2024-11-22)
 
 #### Fixed